requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.13
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
idna==3.7
importlib_metadata==8.0.0
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
requests==2.32.3
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
zipp==3.19.2
